fix_test_trt_fc_fuse_quant_dequant_pass
